requirements.txt
Django==4.2.2
Pillow==10.0.0
django-model-utils==4.3.1
